requirements.txt
# ---------- BASE ----------
streamlit>=1.33
pandas>=2.1
numpy>=1.26
scikit-learn>=1.4
lightgbm>=4.0
shap>=0.45
matplotlib>=3.8
plotly>=5.20
seaborn>=0.13
pyarrow>=15.0
openpyxl>=3.1
joblib>=1.3
tqdm>=4.66

# ---------- TEXT ----------
sentence-transformers>=2.6
scipy>=1.11
nltk>=3.9

# ---------- IMAGE ----------
torch>=2.2
torchvision>=0.17
opencv-python>=4.9
Pillow>=10.0
open-clip-torch>=2.24.0

# ---------- AUDIO ----------
torchaudio>=2.2
librosa>=0.10
soundfile>=0.12

# ---------- VIDEO ----------
decord>=0.6.0

# ---------- RAG ----------
faiss-cpu>=1.8.0

# ---------- DEV ----------
pytest>=8.1
ipykernel>=6.29
notebook>=7.2
black>=24.3
flake8>=7.0
